+++ WalletWasabi.Fluent/ViewModels/StatusIcon/IStatusIconViewModel.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Windows.Input;
 using WalletWasabi.Fluent.AppServices.Tor;
+using WalletWasabi.Services.Tor.StatusChecker;
 
 namespace WalletWasabi.Fluent.ViewModels.StatusIcon;
