+++ WalletWasabi.Fluent/ViewModels/StatusIcon/StatusIconDesignViewModel.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Windows.Input;
 using ReactiveUI;
 using WalletWasabi.Fluent.AppServices.Tor;
+using WalletWasabi.Services.Tor.StatusChecker;
 
 namespace WalletWasabi.Fluent.ViewModels.StatusIcon;
