+++ WalletWasabi.Fluent/Views/StatusIcon/TorIssuesControl.axaml
@@ -3,7 +3,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:statusIcon="clr-namespace:WalletWasabi.Fluent.ViewModels.StatusIcon"
-             xmlns:tor="clr-namespace:WalletWasabi.Fluent.AppServices.Tor"
+             xmlns:tor="clr-namespace:WalletWasabi.Services.Tor.StatusChecker;assembly=WalletWasabi"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="WalletWasabi.Fluent.Views.StatusIcon.TorIssuesControl"
              x:CompileBindings="True"
